Fix the UB that made the debug test suite unrunnable, and what it hid
`store_u8_n` stored four bytes through a `*mut i32`. That pointer is a
row of a picture and is aligned to nothing in particular, so the store
promises an alignment it does not have. x86 tolerates it and release
builds do not check, which is why it survived: the only symptom was that
`cargo test` without `--release` aborted the whole run at
`dsp::h264_avx2::tests::qpel_matches_scalar`, and everybody ran the
release suite.

Two sites, not one. `h264_avx2.rs` is where it was reported;
`h264_x86_128.rs` has the same line, which turned up by grepping for the
pattern rather than fixing the address in the report.

**Running the debug suite for the first time then found two more bugs**,
both in test data generation and both hidden by release wrapping. The
HEVC 8-bit combine test and the AVX-512 idct32 test each build a random
value in [-range, +range] by computing it in i16, where the intermediate
does not fit: at range 22500 and 32767 the cast wraps before the
subtraction. In release that silently produced a different distribution
than intended, so those kernels were being exercised over the wrong
range; in debug it overflows. Both now compute in i32 and cast once.

The debug suite passes: 87 tests. It has never passed before, because
it has never run.

Also anchors `verify.sh`. It cd's to `H26X_WORK` before resolving the
decoder path, so a relative path given from anywhere else resolves to
nothing — and a decoder that cannot run produces empty output for every
fixture and every rung. The fixture check notices, because it compares
against recorded hashes. **The rung sweep does not**: it compares the
rungs against each other, and seven empty outputs agree perfectly. That
is a self-consistency check with no anchor, which is the same failure as
printing a control and never applying it. It now refuses unless the
decoder exists and produces output, and says why.

+++ src/dsp/hevc_avx2_u8.rs
@@ -1261,8 +1261,8 @@ mod tests {
         let max = 255;
         for &(w, h) in &[(2usize, 4usize), (4, 4), (6, 8), (8, 8), (12, 16), (16, 8), (24, 4), (32, 32), (64, 64)] {
             for range in [16000i32, 22500] {
-                let a: Vec<i16> = (0..w * h).map(|_| (lcg(&mut seed) % (2 * range as u32)) as i16 - range as i16).collect();
-                let b: Vec<i16> = (0..w * h).map(|_| (lcg(&mut seed) % (2 * range as u32)) as i16 - range as i16).collect();
+                let a: Vec<i16> = (0..w * h).map(|_| ((lcg(&mut seed) % (2 * range as u32)) as i32 - range) as i16).collect();
+                let b: Vec<i16> = (0..w * h).map(|_| ((lcg(&mut seed) % (2 * range as u32)) as i32 - range) as i16).collect();
                 let stride = w + 5;
                 let mut d1 = vec![0u8; stride * h];
                 let mut d2 = vec![0u8; stride * h];
